New: Major events + profile page restructuring ! Slight changes, code normalisation, events db restructure !!! Changes, removed invisible button replaced with linearlayout !!! Can edit group names and images Unique Groups Done Creating group, can select members, now posts data to firebase server including which groups, needs more testing
