+++ app/src/main/java/example/com/groupeasy/activities/lists_activity.java
@@ -107,6 +107,7 @@ public class lists_activity extends AppCompatActivity {
 
                 }
 
+
                 mAdapter.notifyDataSetChanged();
 
                 if(mLstGroups.isEmpty()){
