+++ app/src/main/java/example/com/groupeasy/pojo/Participants.java
@@ -1,36 +1,36 @@
-//package example.com.groupeasy.pojo;
-//
-///**
-// * Created by Harsh on 23-10-2017.
-// */
-//
-//public class Participants {
-//
-//    String name;
-//    String value;
-//
-//    public Participants(){
-//
-//    }
-//
-//    public Participants(String name, String value) {
-//        this.name = name;
-//        this.value = value;
-//    }
-//
-//    public String getName() {
-//        return name;
-//    }
-//
-//    public void setName(String name) {
-//        this.name = name;
-//    }
-//
-//    public String getValue() {
-//        return value;
-//    }
-//
-//    public void setValue(String value) {
-//        this.value = value;
-//    }
-//}
+package example.com.groupeasy.pojo;
+
+/**
+ * Created by Harsh on 23-10-2017.
+ */
+
+public class Participants {
+
+    String name;
+    String value;
+
+    public Participants(){
+
+    }
+
+    public Participants(String name, String value) {
+        this.name = name;
+        this.value = value;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+}
