+++ app/src/main/java/example/com/groupeasy/pojo/list_primary.java
@@ -14,28 +14,29 @@ public class list_primary {
     String location;
     String id;
 
-    Participants participants;
+//    Participants participants;
 
     public list_primary(){
 
     }
 
-    public list_primary(String name, String admin, String location, String id, Participants participants) {
+//    public list_primary(String name, String admin, String location, String id, Participants participants) {
+    public list_primary(String name, String admin, String location, String id) {
         this.name = name;
         this.admin = admin;
         this.location = location;
         this.id = id;
-        this.participants = participants;
+//        this.participants = participants;
 
     }
 
-    public Participants getParticipants() {
-        return participants;
-    }
-
-    public void setParticipants(Participants participants) {
-        this.participants = participants;
-    }
+//    public Participants getParticipants() {
+//        return participants;
+//    }
+//
+//    public void setParticipants(Participants participants) {
+//        this.participants = participants;
+//    }
 
     public String getId() {
         return id;
@@ -68,34 +69,34 @@ public class list_primary {
         this.location = location;
     }
 
-   public static class Participants{
-
-        String name;
-        String value;
-
-        public Participants(){
-
-        }
-
-        public Participants(String name, String value) {
-            this.name = name;
-            this.value = value;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public String getValue() {
-            return value;
-        }
-
-        public void setValue(String value) {
-            this.value = value;
-        }
-    }
+//   public static class Participants{
+//
+//        String name;
+//        String value;
+//
+//        public Participants(){
+//
+//        }
+//
+//        public Participants(String name, String value) {
+//            this.name = name;
+//            this.value = value;
+//        }
+//
+//        public String getName() {
+//            return name;
+//        }
+//
+//        public void setName(String name) {
+//            this.name = name;
+//        }
+//
+//        public String getValue() {
+//            return value;
+//        }
+//
+//        public void setValue(String value) {
+//            this.value = value;
+//        }
+//    }
 }
