+++ app/src/main/res/drawable/my_circle.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+
+    <stroke android:width="0.5dp" android:color="@color/black" />
+
+    <solid
+        android:color="@color/golden"/>
+
+    <size
+        android:width="25dp"
+        android:height="25dp"/>
+</shape>
