+++ app/src/main/res/drawable/my_rectangle.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/listview_background_shape">
+    <stroke android:width="0.5dp" android:color="@color/black" />
+    <!--<padding android:left="2dp"-->
+        <!--android:top="2dp"-->
+        <!--android:right="2dp"-->
+        <!--android:bottom="2dp" />-->
+    <corners android:radius="1dp" />
+    <solid android:color="@color/red" />
+</shape>
