+++ app/src/main/res/values/colors.xml
@@ -16,6 +16,7 @@
     <color name="cyan">#26c6da</color>
     <color name="cyan_dark">#1ea0b0</color>
     <color name="light_blue">#b9cfe4</color>
+    <color name="purple">#ba68c8</color>
 
 
     <!--taken from v_1-->
